Сhanges due to the transition to the structure MVVM
