+++ RegistrationWindow/RegistrationWindow/Coordinators/Protocols/Coordinators.swift
@@ -1,0 +1,13 @@
+//
+//  Coordinators.swift
+//  RegistrationWindow
+//
+//  Created by MAC on 30.10.2022.
+//
+
+import UIKit
+protocol Coordinator {
+    var navigationController: UINavigationController { get set }
+    
+    func start()
+}
