+++ RegistrationWindow/RegistrationWindow/VIew/RegistrationController.swift
@@ -1,29 +1,39 @@
 import UIKit
 
-class ViewController: UIViewController {
+class RegistController: UIViewController {
     private var textfieldName = UITextField()
     private var textfieldPassword = UITextField()
+    private var acceptButton = UIButton()
     private var infoLabel = UILabel()
-    private var registButton = UIButton()
-    private var loginButton = UIButton()
+    var viewModel: RegistrationViewModel?
+    var coordinator: Coordinator?
+    
+    init(viewModel: RegistrationViewModel) {
+        self.viewModel = viewModel
+        super.init(nibName: nil, bundle: nil)
+    }
+    
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+    }
+
     override func viewDidLoad() {
         super.viewDidLoad()
         setupSubview()
         setupAutoLoyaut()
         setupStyle()
         setupAction()
-        view.backgroundColor = .systemBlue
+        view.backgroundColor = .blue
     }
     private func setupAutoLoyaut() {
         textfieldName.translatesAutoresizingMaskIntoConstraints = false
         textfieldPassword.translatesAutoresizingMaskIntoConstraints = false
         infoLabel.translatesAutoresizingMaskIntoConstraints = false
-        registButton.translatesAutoresizingMaskIntoConstraints = false
-        loginButton.translatesAutoresizingMaskIntoConstraints = false
+        acceptButton.translatesAutoresizingMaskIntoConstraints = false
         NSLayoutConstraint.activate([
             textfieldName.topAnchor.constraint(equalTo:view.centerYAnchor,constant: -50),
             textfieldName.rightAnchor.constraint(equalTo: view.rightAnchor,constant: -100),
-            textfieldName.leftAnchor.constraint(equalTo: view.leftAnchor,constant: 90),
+            textfieldName.widthAnchor.constraint(equalToConstant: 200),
             textfieldName.heightAnchor.constraint(equalToConstant: 40),
             
             textfieldPassword.heightAnchor.constraint(equalToConstant: 40),
@@ -32,19 +42,14 @@ class ViewController: UIViewController {
             textfieldPassword.rightAnchor.constraint(equalTo: textfieldName.rightAnchor),
             
             infoLabel.heightAnchor.constraint(equalToConstant: 60),
-            infoLabel.widthAnchor.constraint(equalToConstant: 140),
+            infoLabel.widthAnchor.constraint(equalToConstant: 250),
             infoLabel.topAnchor.constraint(equalTo: textfieldName.bottomAnchor, constant: -150),
-            infoLabel.rightAnchor.constraint(equalTo: textfieldName.rightAnchor,constant: -15),
+            infoLabel.leftAnchor.constraint(equalTo: textfieldName.leftAnchor),
             
-            registButton.widthAnchor.constraint(equalToConstant: 100),
-            registButton.heightAnchor.constraint(equalToConstant: 30),
-            registButton.topAnchor.constraint(equalTo: textfieldName.topAnchor, constant: 130),
-            registButton.rightAnchor.constraint(equalTo: textfieldName.rightAnchor,constant: -50),
-            
-            loginButton.widthAnchor.constraint(equalToConstant: 100),
-            loginButton.heightAnchor.constraint(equalToConstant: 30),
-            loginButton.topAnchor.constraint(equalTo: textfieldName.topAnchor, constant: 160),
-            loginButton.rightAnchor.constraint(equalTo: textfieldName.rightAnchor,constant: -50)
+            acceptButton.widthAnchor.constraint(equalToConstant: 100),
+            acceptButton.heightAnchor.constraint(equalToConstant: 30),
+            acceptButton.topAnchor.constraint(equalTo: textfieldName.topAnchor, constant: 130),
+            acceptButton.rightAnchor.constraint(equalTo: textfieldName.rightAnchor,constant: -50)
         ])
     }
     
@@ -52,40 +57,35 @@ class ViewController: UIViewController {
         view.addSubview(textfieldName)
         view.addSubview(textfieldPassword)
         view.addSubview(infoLabel)
-        view.addSubview(registButton)
-        view.addSubview(loginButton)
+        view.addSubview(acceptButton)
     }
     private func setupStyle() {
         textfieldName.borderStyle = UITextField.BorderStyle.bezel
-        textfieldName.placeholder = "press your name"
+        textfieldName.placeholder = "Create name"
         textfieldName.backgroundColor = UIColor.white
         
         textfieldPassword.borderStyle = UITextField.BorderStyle.bezel
-        textfieldPassword.placeholder = "press your password"
+        textfieldPassword.placeholder = "Create password"
         textfieldPassword.backgroundColor = UIColor.white
         
         infoLabel.textColor = .yellow
         infoLabel.font = UIFont.systemFont(ofSize: 40)
-        infoLabel.text = "Hello!"
+        infoLabel.text = "Registration"
         
-        registButton.backgroundColor = .systemBlue
-        registButton.setTitle("Registration", for: .normal)
-        registButton.setTitleColor(.yellow, for: .normal)
-        
-        loginButton.setTitle("Ok", for: .normal)
-        loginButton.setTitleColor(.yellow, for: .normal)
-        loginButton.backgroundColor = .systemBlue
+        acceptButton.setTitle("Save", for: .normal)
+        acceptButton.setTitleColor(.yellow, for: .normal)
+        acceptButton.backgroundColor = .systemBlue
     }
     
     override func viewWillAppear(_ animated: Bool) {
-        self.navigationController?.setNavigationBarHidden(true, animated: false)
+        self.navigationController?.setNavigationBarHidden(false, animated: false)
+        navigationController?.navigationBar.barTintColor = UIColor.blue
     }
-    
+        
     private func setupAction() {
         textfieldName.addTarget(self, action: #selector(hideKeyboard), for: .editingDidEndOnExit)
         textfieldPassword.addTarget(self, action: #selector(passwordChanged), for: .editingDidEndOnExit)
-        registButton.addTarget(self, action: #selector(tapRegist), for: .touchUpInside)
-        loginButton.addTarget(self, action: #selector(tapLogin), for: .touchUpInside)
+        acceptButton.addTarget(self, action: #selector(saveButton), for: .touchUpInside)
     }
     
     @objc func hideKeyboard() {
@@ -95,13 +95,11 @@ class ViewController: UIViewController {
     @objc func passwordChanged() {
     }
     
-    @objc func tapRegist() {
-        let regVC = RegistController()
-        navigationController?.pushViewController(regVC, animated: true)
-    }
-    
-    @objc func tapLogin() {
-            let logVC = FinalViewController()
-            navigationController?.pushViewController(logVC, animated: true)
+    @objc func saveButton() {
+        navigationController?.popViewController(animated: true)
+        let nameText = textfieldName.text?.trimmingCharacters(in: .whitespaces)
+        let passwordText = textfieldPassword.text?.trimmingCharacters(in: .whitespaces)
+        viewModel?.model?.name = nameText
+        viewModel?.model?.password = passwordText
     }
 }
