+++ RegistrationWindow/RegistrationWindow/VIew/WeatherViewController.swift
@@ -1,7 +1,18 @@
 import UIKit
 
-class FinalViewController: UIViewController {
+class WeatherViewController: UIViewController {
     private let greetingsLabel = UILabel()
+    var viewModel: WeatherViewModel?
+    var coordinator: AppCoordinator?
+    
+    init(viewModel: WeatherViewModel) {
+        self.viewModel = viewModel
+        super.init(nibName: nil, bundle: nil)
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .blue
@@ -13,7 +24,7 @@ class FinalViewController: UIViewController {
         greetingsLabel.translatesAutoresizingMaskIntoConstraints = false
         NSLayoutConstraint.activate([
             greetingsLabel.topAnchor.constraint(equalTo: view.topAnchor,constant: 200),
-            greetingsLabel.leftAnchor.constraint(equalTo: view.leftAnchor, constant: 40),
+            greetingsLabel.leftAnchor.constraint(equalTo: view.leftAnchor, constant: 60),
             greetingsLabel.widthAnchor.constraint(equalToConstant: 400),
             greetingsLabel.heightAnchor.constraint(equalToConstant: 60)
         ])
@@ -30,6 +41,8 @@ class FinalViewController: UIViewController {
         greetingsLabel.alpha = 0
         UIView.animate(withDuration: 2) {
             self.greetingsLabel.alpha = 1
+            self.greetingsLabel.transform = CGAffineTransform(scaleX: 1.2, y: 1.2)
+            self.greetingsLabel.textColor = .systemOrange
         }
     }
