+++ RegistrationWindow/RegistrationWindow/ViewModel/RegistrationViewModel.swift
@@ -1,0 +1,16 @@
+//
+//  RegistrationViewModel.swift
+//  RegistrationWindow
+//
+//  Created by MAC on 30.10.2022.
+//
+
+import UIKit
+
+class RegistrationViewModel {
+    var model: UserModel?
+    
+    init(model: UserModel) {
+        self.model = model
+    }
+}
