+++ RegistrationWindow/RegistrationWindow/ViewModel/WeatherViewModel.swift
@@ -1,0 +1,17 @@
+//
+//  FinalViewModel.swift
+//  RegistrationWindow
+//
+//  Created by MAC on 30.10.2022.
+//
+
+import UIKit
+class WeatherViewModel {
+    var model: UserModel?
+    
+    init(model: UserModel) {
+        self.model = model
+    }
+    
+    
+}
